adding file path for image to profiles database
